@@ -1,50 +1,57 @@
 {
-	"name": "angular-tour-of-heroes",
+	"name": "angular-io-example",
 	"version": "1.0.0",
-	"license": "MIT",
+	"private": true,
+	"description": "Example project from an angular.io guide.",
 	"scripts": {
 		"ng": "ng",
-		"start": "ng serve",
 		"build": "ng build",
+		"start": "ng serve",
 		"test": "ng test",
-		"lint": "ng lint",
+		"lint": "tslint ./src/**/*.ts -t verbose",
 		"e2e": "ng e2e"
 	},
-	"private": true,
+	"keywords": [],
+	"author": "",
+	"license": "MIT",
 	"dependencies": {
-		"@angular/animations": "^4.0.0",
-		"@angular/common": "^4.0.0",
-		"@angular/compiler": "^4.0.0",
-		"@angular/core": "^4.0.0",
-		"@angular/forms": "^4.0.0",
-		"@angular/http": "^4.0.0",
-		"@angular/platform-browser": "^4.0.0",
-		"@angular/platform-browser-dynamic": "^4.0.0",
-		"@angular/router": "^4.0.0",
+		"@angular/animations": "~5.2.2",
+		"@angular/common": "~5.2.2",
+		"@angular/compiler": "~5.2.2",
+		"@angular/compiler-cli": "~5.2.2",
+		"@angular/core": "~5.2.2",
+		"@angular/forms": "~5.2.2",
+		"@angular/http": "~5.2.2",
+		"@angular/platform-browser": "~5.2.2",
+		"@angular/platform-browser-dynamic": "~5.2.2",
+		"@angular/platform-server": "~5.2.2",
+		"@angular/router": "~5.2.2",
+		"@angular/upgrade": "~5.2.2",
+		"angular-in-memory-web-api": "~0.5.0",
 		"core-js": "^2.4.1",
-		"rxjs": "^5.1.0",
-		"zone.js": "^0.8.4",
-		"angular-in-memory-web-api": "^0.3.1"
+		"rxjs": "^5.5.0",
+		"web-animations-js": "^2.3.1",
+		"zone.js": "^0.8.4"
 	},
 	"devDependencies": {
-		"@angular/cli": "1.2.0",
-		"@angular/compiler-cli": "^4.0.0",
-		"@angular/language-service": "^4.0.0",
-		"@types/jasmine": "~2.5.53",
-		"@types/jasminewd2": "~2.0.2",
-		"@types/node": "~6.0.60",
-		"codelyzer": "~3.0.1",
-		"jasmine-core": "~2.6.2",
-		"jasmine-spec-reporter": "~4.1.0",
-		"karma": "~1.7.0",
-		"karma-chrome-launcher": "~2.1.1",
-		"karma-cli": "~1.0.1",
-		"karma-coverage-istanbul-reporter": "^1.2.1",
-		"karma-jasmine": "~1.1.0",
+		"@angular/cli": "1.6.6",
+		"@types/jasmine": "~2.8.6",
+		"@types/jasminewd2": "^2.0.3",
+		"@types/node": "^9.4.0",
+		"jasmine": "~2.9.0",
+		"jasmine-core": "~2.9.1",
+		"karma": "^2.0.0",
+		"karma-chrome-launcher": "^2.0.0",
+		"karma-cli": "^1.0.1",
+		"karma-jasmine": "^1.0.2",
 		"karma-jasmine-html-reporter": "^0.2.2",
-		"protractor": "~5.1.2",
-		"ts-node": "~3.0.4",
-		"tslint": "~5.3.2",
-		"typescript": "~2.3.3"
-	}
+		"karma-phantomjs-launcher": "^1.0.2",
+		"lodash": "^4.16.2",
+		"phantomjs-prebuilt": "^2.1.7",
+		"protractor": "~5.3.0",
+		"ts-node": "^4.1.0",
+		"tslint": "^5.9.1",
+		"typescript": "2.6.2"
+	},
+	"repository": {}
 }
