@@ -1,35 +1,45 @@
-import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
+import { BrowserModule } from '@angular/platform-browser';
 import { FormsModule } from '@angular/forms';
-import { HttpModule } from '@angular/http';
+import { HttpClientModule } from '@angular/common/http';
 
-import { InMemoryWebApiModule } from 'angular-in-memory-web-api';
+import { HttpClientInMemoryWebApiModule } from 'angular-in-memory-web-api';
 import { InMemoryDataService } from './in-memory-data.service';
 
-import { AppComponent } from './app.component';
 import { AppRoutingModule } from './app-routing.module';
+
+import { AppComponent } from './app.component';
+import { DashboardComponent } from './dashboard/dashboard.component';
+import { HeroDetailComponent } from './hero-detail/hero-detail.component';
+import { HeroesComponent } from './heroes/heroes.component';
+import { HeroSearchComponent } from './hero-search/hero-search.component';
 import { HeroService } from './hero.service';
-import { DashboardComponent } from './dashboard.component';
-import { HeroesComponent } from './heroes.component';
-import { HeroDetailComponent } from './hero-detail.component';
-import { HeroSearchComponent } from './hero-search.component';
+import { MessageService } from './message.service';
+import { MessagesComponent } from './messages/messages.component';
 
 @NgModule({
 	imports: [
 		BrowserModule,
 		FormsModule,
 		AppRoutingModule,
-		HttpModule,
-		InMemoryWebApiModule.forRoot(InMemoryDataService, { delay: 600 })
+		HttpClientModule,
+
+		// The HttpClientInMemoryWebApiModule module intercepts HTTP requests
+		// and returns simulated server responses.
+		// Remove it when a real server is ready to receive requests.
+		HttpClientInMemoryWebApiModule.forRoot(InMemoryDataService, {
+			dataEncapsulation: false
+		})
 	],
 	declarations: [
 		AppComponent,
 		DashboardComponent,
-		HeroSearchComponent,
 		HeroesComponent,
-		HeroDetailComponent
+		HeroDetailComponent,
+		MessagesComponent,
+		HeroSearchComponent
 	],
-	providers: [HeroService],
+	providers: [HeroService, MessageService],
 	bootstrap: [AppComponent]
 })
 export class AppModule {}
