@@ -1,78 +1,131 @@
 import { Injectable } from '@angular/core';
-import { Headers, Http, Response } from '@angular/http';
+import { HttpClient, HttpHeaders } from '@angular/common/http';
 
-import 'rxjs/add/operator/toPromise';
+import { Observable } from 'rxjs/Observable';
+import { of } from 'rxjs/observable/of';
+import { catchError, map, tap } from 'rxjs/operators';
 
 import { Hero } from './hero';
+import { MessageService } from './message.service';
+
+const httpOptions = {
+	headers: new HttpHeaders({ 'Content-Type': 'application/json' })
+};
 
 @Injectable()
 export class HeroService {
-	private heroesUrl = 'app/heroes'; // URL to web api
+	private heroesUrl = 'api/heroes';
+	// URL to web api
 
-	constructor(private http: Http) {}
+	constructor(
+		private http: HttpClient,
+		private messageService: MessageService
+	) {}
 
-	getHeroes(): Promise<Array<Hero>> {
+	/** GET heroes from the server */
+	getHeroes(): Observable<Hero[]> {
 		return this.http
-			.get(this.heroesUrl)
-			.toPromise()
-			.then(response => {
-				return response.json().data as Hero[];
-			})
-			.catch(this.handleError);
+			.get<Hero[]>(this.heroesUrl)
+			.pipe(
+				tap(heroes => this.log(`fetched heroes`)),
+				catchError(this.handleError('getHeroes', []))
+			);
 	}
 
-	getHero(id: number): Promise<Hero> {
-		return this.getHeroes().then(heroes => heroes.find(hero => hero.id === id));
+	/** GET hero by id. Return `undefined` when id not found */
+	getHeroNo404<Data>(id: number): Observable<Hero> {
+		const url = `${this.heroesUrl}/?id=${id}`;
+		return this.http.get<Hero[]>(url).pipe(
+			map(heroes => heroes[0]), // returns a {0|1} element array
+			tap(h => {
+				const outcome = h ? `fetched` : `did not find`;
+				this.log(`${outcome} hero id=${id}`);
+			}),
+			catchError(this.handleError<Hero>(`getHero id=${id}`))
+		);
 	}
 
-	save(hero: Hero): Promise<Hero> {
-		if (hero.id) {
-			return this.put(hero);
+	/** GET hero by id. Will 404 if id not found */
+	getHero(id: number): Observable<Hero> {
+		const url = `${this.heroesUrl}/${id}`;
+		return this.http
+			.get<Hero>(url)
+			.pipe(
+				tap(_ => this.log(`fetched hero id=${id}`)),
+				catchError(this.handleError<Hero>(`getHero id=${id}`))
+			);
+	}
+
+	/* GET heroes whose name contains search term */
+	searchHeroes(term: string): Observable<Hero[]> {
+		if (!term.trim()) {
+			// if not search term, return empty hero array.
+			return of([]);
 		}
-		return this.post(hero);
+		return this.http
+			.get<Hero[]>(`api/heroes/?name=${term}`)
+			.pipe(
+				tap(_ => this.log(`found heroes matching "${term}"`)),
+				catchError(this.handleError<Hero[]>('searchHeroes', []))
+			);
 	}
 
-	delete(hero: Hero): Promise<Response> {
-		const headers = new Headers();
-		headers.append('Content-Type', 'application/json');
+	//////// Save methods //////////
 
-		const url = `${this.heroesUrl}/${hero.id}`;
+	/** POST: add a new hero to the server */
+	addHero(hero: Hero): Observable<Hero> {
+		return this.http
+			.post<Hero>(this.heroesUrl, hero, httpOptions)
+			.pipe(
+				tap((hero: Hero) => this.log(`added hero w/ id=${hero.id}`)),
+				catchError(this.handleError<Hero>('addHero'))
+			);
+	}
+
+	/** DELETE: delete the hero from the server */
+	deleteHero(hero: Hero | number): Observable<Hero> {
+		const id = typeof hero === 'number' ? hero : hero.id;
+		const url = `${this.heroesUrl}/${id}`;
 
 		return this.http
-			.delete(url, { headers: headers })
-			.toPromise()
-			.catch(this.handleError);
+			.delete<Hero>(url, httpOptions)
+			.pipe(
+				tap(_ => this.log(`deleted hero id=${id}`)),
+				catchError(this.handleError<Hero>('deleteHero'))
+			);
 	}
 
-	// Add new Hero
-	private post(hero: Hero): Promise<Hero> {
-		const headers = new Headers({
-			'Content-Type': 'application/json'
-		});
-
+	/** PUT: update the hero on the server */
+	updateHero(hero: Hero): Observable<any> {
 		return this.http
-			.post(this.heroesUrl, JSON.stringify(hero), { headers: headers })
-			.toPromise()
-			.then(res => res.json().data)
-			.catch(this.handleError);
+			.put(this.heroesUrl, hero, httpOptions)
+			.pipe(
+				tap(_ => this.log(`updated hero id=${hero.id}`)),
+				catchError(this.handleError<any>('updateHero'))
+			);
 	}
 
-	// Update existing Hero
-	private put(hero: Hero): Promise<Hero> {
-		const headers = new Headers();
-		headers.append('Content-Type', 'application/json');
+	/**
+	 * Handle Http operation that failed.
+	 * Let the app continue.
+	 * @param operation - name of the operation that failed
+	 * @param result - optional value to return as the observable result
+	 */
+	private handleError<T>(operation = 'operation', result?: T) {
+		return (error: any): Observable<T> => {
+			// TODO: send the error to remote logging infrastructure
+			console.error(error); // log to console instead
 
-		const url = `${this.heroesUrl}/${hero.id}`;
+			// TODO: better job of transforming error for user consumption
+			this.log(`${operation} failed: ${error.message}`);
 
-		return this.http
-			.put(url, JSON.stringify(hero), { headers: headers })
-			.toPromise()
-			.then(() => hero)
-			.catch(this.handleError);
+			// Let the app keep running by returning an empty result.
+			return of(result as T);
+		};
 	}
 
-	private handleError(error: any): Promise<any> {
-		console.error('An error occurred', error);
-		return Promise.reject(error.message || error);
+	/** Log a HeroService message with the MessageService */
+	private log(message: string) {
+		this.messageService.add('HeroService: ' + message);
 	}
 }
