@@ -1,11 +1,11 @@
-import { browser, element, by } from 'protractor';
+import { browser, by, element } from 'protractor';
 
-export class BlankPage {
+export class AppPage {
 	navigateTo() {
 		return browser.get('/');
 	}
 
 	getParagraphText() {
-		return element(by.css('my-root h1')).getText();
+		return element(by.css('app-root h1')).getText();
 	}
 }
