@@ -7,7 +7,11 @@ exports.config = {
 	allScriptsTimeout: 11000,
 	specs: ['./e2e/**/*.e2e-spec.ts'],
 	capabilities: {
-		browserName: 'chrome'
+		browserName: 'chrome',
+		// For Travis CI only
+		chromeOptions: {
+			binary: process.env.CHROME_BIN
+		}
 	},
 	directConnect: true,
 	baseUrl: 'http://localhost:4200/',
