@@ -1,16 +1,8 @@
 import { Component } from '@angular/core';
 
 @Component({
-	selector: 'my-root',
-	template: `
-    <h1>{{title}}</h1>
-    <div class="header-bar"></div>
-    <nav>
-      <a routerLink="/dashboard" routerLinkActive="active">Dashboard</a>
-      <a routerLink="/heroes" routerLinkActive="active">Heroes</a>
-    </nav>
-    <router-outlet></router-outlet>
-  `,
+	selector: 'app-root',
+	templateUrl: './app.component.html',
 	styleUrls: ['./app.component.css']
 })
 export class AppComponent {
