@@ -1,4 +1,6 @@
-export class InMemoryDataService {
+import { InMemoryDbService } from 'angular-in-memory-web-api';
+
+export class InMemoryDataService implements InMemoryDbService {
 	createDb() {
 		const heroes = [
 			{ id: 11, name: 'Mr. Nice' },
